oauth/requirements.txt
django==1.11
markdown==2.6.10
django-markup==1.2
django-cors-headers==2.1.0
django-oauth-toolkit==1.0.0
mysqlclient
django-pure-pagination==0.3.0
django-simple-captcha==0.5.6
django-bootstrap3==9.1.0
django-registration-redux==2.0
django-mailgun==0.9.1
pytz
psycopg2

